% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.grid.R
\name{adj.grid}
\alias{adj.grid}
\title{Find adjacency matrix}
\usage{
adj.grid(coords, order = 1, spam = F)
}
\arguments{
\item{coords}{data frame of x, y coordinates}

\item{order}{order of neighborhood}

\item{spam}{save as sparse matrix?}
}
\value{
Adjacency matrix
}
\description{
Find adjacency matrix
}
