% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfdr.R
\name{BFDRhat}
\alias{BFDRhat}
\title{bfdr}
\usage{
BFDRhat(pi1, cutoff)
}
\arguments{
\item{pi1}{adsfa}

\item{cutoff}{adf}

\item{pi0}{adsfa}

\item{cutoff2}{adf}

\item{pi0k}{adfadf}
}
\value{
qval
}
\description{
bfdr
}
