% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casseseModelFit_noTestingTry.R
\name{cass}
\alias{cass}
\title{model fitting cassese et al}
\usage{
cass(msset, roiFactor, logscale = T, thresholds = 1:10)
}
\arguments{
\item{msset}{adfaf}

\item{roiFactor}{adfadf}

\item{logscale}{asdfa}

\item{thresholds}{adfdf}
}
\value{
list
}
\description{
model fitting cassese et al
}
