% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMSI_baseline.R
\name{compareMSI2}
\alias{compareMSI2}
\title{fit model}
\usage{
compareMSI2(msset, conditionOfInterest, feature, nsim = 5000, burnin = 2500,
  trace = T, piPrior = 0.1, seed = 1, logbase2 = F, coord = NULL,
  type.neighbor = "radius", radius.neighbor = 1, maxdist.neighbor = NULL,
  spInit = NULL, bioRep = NULL, techRep)
}
\arguments{
\item{msset}{adfdf}
}
\value{
res
}
\description{
fit model
}
