% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMSI_baselineCenter.R
\name{compareMSI3}
\alias{compareMSI3}
\title{fit model}
\usage{
compareMSI3(msset, conditionOfInterest, feature, nsim = 5000, burnin = 2500,
  trace = T, piPrior = 0.1, seed = 1, logbase2 = F, coord = NULL,
  type.neighbor = "radius", radius.neighbor = 1, maxdist.neighbor = NULL,
  spInit = NULL, bioRep = NULL, techRep, beta0 = 0, prec0 = 0.01,
  precAlpha0 = 0.01, d0 = 0.001, g0 = 0.001, rd = 1e-04)
}
\arguments{
\item{msset}{adfdf}
}
\value{
res
}
\description{
fit model
}
