% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMSI_zeros.R
\name{compareMSI_zeros}
\alias{compareMSI_zeros}
\title{Fit hiearchical spatial model to MSI data, with improved mixing}
\usage{
compareMSI_zeros(msset, conditionOfInterest, feature, nsim = 5000,
  burnin = 2500, trace = T, piPrior = 0.1, seed = 1, logbase2 = F,
  coord = NULL, type.neighbor = "radius", radius.neighbor = 1,
  maxdist.neighbor = NULL, spInit = NULL, bioRep = NULL, techRep,
  beta0 = 0, prec0 = 0.01, precAlpha0 = 0.01, d0 = 0.001, g0 = 0.001,
  rd = 1e-05, dropZeros = T)
}
\arguments{
\item{msset}{an object of class "MSImageSet"}

\item{conditionOfInterest}{a vector or factor giving the level of the condition of interest for each pixel in msset}

\item{feature}{the index of the m/z features for which the model should be fit}

\item{nsim}{number of desired MCMC samples}

\item{burnin}{number of MCMC samples to discard}

\item{trace}{logical, should the full list of MCMC samples be returned for each variable?}

\item{piPrior}{prior probability of differential abundance}

\item{seed}{random seed}

\item{logbase2}{logical, should the intensities be log transformed?}

\item{coord}{data fram of coordinates of the MSImageSet, with columns 'x' and 'y'}

\item{type.neighbor}{neighborhood type (see adj.grid)}

\item{radius.neighbor}{desired neighborhood radius if neighborhood type 'radius' is selected (see adj.grid)}

\item{maxdist.neighbor}{maximum distance for locations to be considered neighbors if neighborhood type 'max.dist' is selected (see adj.grid)}

\item{spInit}{optional, provide precomputed spatial information from output of intializeSpatial}

\item{bioRep}{optional, vector or factor giving the individual/donor to which pixel in the msset belongs}

\item{techRep}{vector or factor giving the tissue to which each pixel in the msset belongs}

\item{beta0}{prior mean of baseline effect}

\item{prec0}{prior variance of baseline effect}

\item{precAlpha0}{prior mean of condition 2 effect}

\item{d0}{shape parameter of hyperprior of variances}

\item{g0}{scale parameter of hyperprior of variances}

\item{rd}{ratio of spike variance to slab variance for condition 2 effect}
}
\value{
res
}
\description{
compareMSI_hc is used to fit a hiearchical Bayesian spatial model to MSI data using a Gibbs Sampler MCMC approach. The model is fit separately for each m/z feature. This version is distinct from compareMSI because it uses hiearchical centering to improve mixing of the MCMC. This function will be called by msiCompare if the MSImageSet is a multi tissue experiment with all tissues coming from the same donor or all from different donors.
}
