% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMSI_zerosSingle.R
\name{compareMSI_zerosSingle}
\alias{compareMSI_zerosSingle}
\title{Fit hiearchical spatial model to MSI data}
\usage{
compareMSI_zerosSingle(msset, conditionOfInterest, feature, nsim = 5000,
  burnin = 2500, trace = T, piPrior = 0.1, seed = 1, logbase2 = F,
  coord = NULL, type.neighbor = "radius", radius.neighbor = 1,
  maxdist.neighbor = NULL, spInit = NULL, bioRep = NULL, techRep,
  beta0 = 0, prec0 = 0.01, precAlpha0 = 0.01, a0_eps = 0.001,
  b0_eps = 0.001, a0_bio = 0.001, b0_bio = 0.001, a0_tec = 0.001,
  b0_tec = 0.001, a0_sp = 0.001, b0_sp = 0.001, rd = 1e-05,
  dropZeros = T)
}
\arguments{
\item{msset}{an object of class "MSImageSet"}

\item{conditionOfInterest}{a vector or factor giving the level of the condition of interest for each pixel in msset}

\item{feature}{the index of the m/z features for which the model should be fit}

\item{nsim}{number of desired MCMC samples}

\item{burnin}{number of MCMC samples to discard}

\item{trace}{logical, should the full list of MCMC samples be returned for each variable?}

\item{piPrior}{prior probability of differential abundance}

\item{seed}{random seed}

\item{logbase2}{logical, should the intensities be log transformed?}

\item{coord}{data fram of coordinates of the MSImageSet, with columns 'x' and 'y'}

\item{type.neighbor}{neighborhood type (see adj.grid)}

\item{radius.neighbor}{desired neighborhood radius if neighborhood type 'radius' is selected (see adj.grid)}

\item{maxdist.neighbor}{maximum distance for locations to be considered neighbors if neighborhood type 'max.dist' is selected (see adj.grid)}

\item{spInit}{optional, provide precomputed spatial information from output of intializeSpatial}

\item{bioRep}{optional, vector or factor giving the individual/donor to which pixel in the msset belongs}

\item{techRep}{vector or factor giving the tissue to which each pixel in the msset belongs}

\item{beta0}{prior mean of baseline effect}

\item{prec0}{prior variance of baseline effect}

\item{precAlpha0}{prior mean of condition 2 effect}

\item{a0_eps}{shape parameter for measurment error precision hyperprior}

\item{b0_eps}{scale parameter for measurment error precision hyperprior}

\item{a0_bio}{shape parameter for biological replicate error precision hyperprior}

\item{b0_bio}{scale parameter for biological replicate error precision hyperprior}

\item{a0_tec}{shape parameter for sample to sample error precision hyperprior}

\item{b0_tec}{scale parameter for sample to sample error precision hyperprior}

\item{a0_sp}{shape parameter for spatial precision hyperprior}

\item{b0_sp}{scale parameter for spatial precision hyperprior}

\item{rd}{ratio of spike variance to slab variance for condition 2 effect}
}
\value{
res
}
\description{
compareMSI is used to fit a hiearchical Bayesian spatial model to MSI data using a Gibbs Sampler MCMC approach. The model is fit separately for each m/z feature.
}
