% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSpatial.R
\name{initializeSpatial}
\alias{initializeSpatial}
\title{initSpatial}
\usage{
initializeSpatial(conditionNames, conditionOfInterest, coord, type.neighbor,
  radius.neighbor, maxdist.neighbor, nsl)
}
\arguments{
\item{conditionNames}{condnames}
}
\value{
init
}
\description{
init
}
