% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_averaging_allVars.R
\name{p_averaging}
\alias{p_averaging}
\title{t-test on averages}
\usage{
p_averaging(simSet)
}
\arguments{
\item{simSet}{An MSImageSet object.}
}
\value{
The estimate of tissue-to-tissue variance, overall mean, estimate of condition difference, and p-value for the test.
}
\description{
T-test on sample averages.
}
\details{
T-test on sample averages. A separate average is produced for each combination of tissue and condition.
}
