% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSim.R
\name{simMulti}
\alias{simMulti}
\title{sim multiple images}
\usage{
simMulti(sampleVar = 0.05, reps = 1, diff = log2(1.5), tau2 = 1,
  sig2 = 0.1, seed = 100, size1 = 7, size2 = size1^2, numHealthy = 2,
  numDisease = 2)
}
\arguments{
\item{sampleVar}{asdfadsf}
}
\value{
simSet
}
\description{
sim multiple images
}
