% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_continuousSlab.R
\name{spatialComparison}
\alias{spatialComparison}
\title{fit model}
\usage{
spatialComparison(msset, sample, conditionOfInterest, feature, nsim = 5000,
  burnin = 2500, trace = T, piPrior = 0.1, seed = 1, logbase2 = F)
}
\arguments{
\item{msset}{adfdf}
}
\value{
res
}
\description{
fit model
}
