% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_continuousSlab2.R
\name{spatialComparison2}
\alias{spatialComparison2}
\title{fit model}
\usage{
spatialComparison2(msset, sample, conditionOfInterest, feature, nsim = 5000,
  burnin = 2500, trace = T, piPrior = 0.1, seed = 1, logbase2 = F,
  coord = NULL, type.neighbor = "radius", radius.neighbor = 1,
  maxdist.neighbor = NULL)
}
\arguments{
\item{msset}{adfdf}
}
\value{
res
}
\description{
fit model
}
