% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_continuousSlab_condTau2nest.R
\name{spatialComparison_condT2nest}
\alias{spatialComparison_condT2nest}
\title{fit model}
\usage{
spatialComparison_condT2nest(msset, conditionOfInterest, feature, nsim = 5000,
  burnin = 2500, trace = T, piPrior = 0.1, seed = 1, logbase2 = F,
  coord = NULL, type.neighbor = "radius", radius.neighbor = 1,
  maxdist.neighbor = NULL, spInit = NULL, bioRep = NULL, techRep)
}
\arguments{
\item{msset}{adfdf}
}
\value{
res
}
\description{
fit model
}
