% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_averaging_allVars.R
\name{tissueWiseANOVA}
\alias{tissueWiseANOVA}
\title{t-test on averages}
\usage{
tissueWiseANOVA(msset, setSamp = pData(msset)$sample,
  setCond = pData(msset)$diagnosis, setBioRep = NULL, logbase2 = F,
  statusUpdates = T)
}
\arguments{
\item{msset}{An MSImageSet object.}

\item{setSamp}{sample labels for msset}

\item{setCond}{condition of interest labels for msset}

\item{setBioRep}{(optional) biological replicate labels for msset}
}
\value{
The estimate(s) of variance, baseline, condition difference, and p-value for the test.
}
\description{
T-test on sample averages.
}
\details{
T-test on sample averages. A separate average is produced for each combination of tissue and condition.
}
