% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSpatial.R
\name{updateSpatial}
\alias{updateSpatial}
\title{update spatial vector}
\usage{
updateSpatial(Wtrip, Wbegfin, m, nsl, phiVec, tau2, rho = 1, eps_m.var,
  offset.phi, tauVar.a, tauVar.b, proper = F)
}
\arguments{
\item{Wtrip}{adfadfa}
}
\value{
phiVec
}
\description{
update spatial vector
}
