% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSpatial_condT2.R
\name{updateSpatial_condT2}
\alias{updateSpatial_condT2}
\title{update spatial vector}
\usage{
updateSpatial_condT2(Wtrip, Wbegfin, m, nsl, phiVec, tau2, rho = 1, eps_m.var,
  offset.phi, tauVar.a, tauVar.b, proper = F, sample, islands = NULL)
}
\arguments{
\item{Wtrip}{adfadfa}
}
\value{
phiVec
}
\description{
update spatial vector
}
