% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSpatial_singeT2.R
\name{updateSpatial_sing}
\alias{updateSpatial_sing}
\title{update spatial vector}
\usage{
updateSpatial_sing(Wtrip, Wbegfin, m, nsl, phiVec, tau2, rho = 1, eps_m.var,
  offset.phi, tauVar.a, tauVar.b, proper = F, sample)
}
\arguments{
\item{Wtrip}{adfadfa}
}
\value{
phiVec
}
\description{
update spatial vector
}
