% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.grid.R
\name{adj.grid}
\alias{adj.grid}
\title{Find adjacency matrix}
\usage{
adj.grid(coords, max.dist = NULL, radius = 1, spam = F, type = "radius",
  sample)
}
\arguments{
\item{coords}{data frame of x, y coordinates}

\item{spam}{save as sparse matrix?}

\item{order}{order of neighborhood}
}
\value{
Adjacency matrix
}
\description{
Find adjacency matrix
}
